refactor(utils): rename get.tsx → link.ts
